Make the re-onboarding guard assertion actually run
2028 is an integer, and complete_student_onboarding takes smallint. With every
other argument untyped there was no candidate to resolve against, so the call
failed with 42883 before it reached the guard. The suite had never been run —
there was no database to run it against — so the one assertion covering
"an onboarded student cannot rerun onboarding and reset a reviewed membership"
had never executed.

With the cast it passes: the guard is real. All 334 assertions across 12 suites
now pass against the local stack.

+++ supabase/tests/membership_review.sql
@@ -125,8 +125,11 @@ select extensions.is(
 set local role authenticated;
 select set_config('request.jwt.claim.sub', '71717171-7171-4171-8171-717171717102', true);
 select extensions.throws_ok(
+  -- 2028 without the cast is an integer, and with every other argument
+  -- untyped there is no candidate to resolve against smallint. The call then
+  -- failed with 42883 before it ever reached the guard this asserts.
   $$select * from public.complete_student_onboarding(
-    'Student', 'B.Tech', 2028, (select id from public.universities where slug = 'bennett-university'),
+    'Student', 'B.Tech', 2028::smallint, (select id from public.universities where slug = 'bennett-university'),
     'ace_exams', 'solo')$$,
   '42501', 'Onboarding is already complete', 're-onboarding cannot reset a reviewed membership'
 );
